+++ flask_dance/contrib/authentiq.py
@@ -1,7 +1,5 @@
-from functools import partial
-
-from flask import _app_ctx_stack as stack
-from flask.globals import LocalProxy, _lookup_app_object
+from flask import g
+from werkzeug.local import LocalProxy
 
 from flask_dance.consumer import OAuth2ConsumerBlueprint
 
@@ -78,10 +76,9 @@ def make_authentiq_blueprint(
 
     @authentiq_bp.before_app_request
     def set_applocal_session():
-        ctx = stack.top
-        ctx.authentiq_oauth = authentiq_bp.session
+        g.flask_dance_authentiq = authentiq_bp.session
 
     return authentiq_bp
 
 
-authentiq = LocalProxy(partial(_lookup_app_object, "authentiq_oauth"))
+authentiq = LocalProxy(lambda: g.flask_dance_authentiq)