+++ flask_dance/contrib/fitbit.py
@@ -1,7 +1,5 @@
-from functools import partial
-
-from flask import _app_ctx_stack as stack
-from flask.globals import LocalProxy, _lookup_app_object
+from flask import g
+from werkzeug.local import LocalProxy
 
 from flask_dance.consumer import OAuth2ConsumerBlueprint
 
@@ -76,10 +74,9 @@ def make_fitbit_blueprint(
 
     @fitbit_bp.before_app_request
     def set_applocal_session():
-        ctx = stack.top
-        ctx.fitbit_oauth = fitbit_bp.session
+        g.flask_dance_fitbit = fitbit_bp.session
 
     return fitbit_bp
 
 
-fitbit = LocalProxy(partial(_lookup_app_object, "fitbit_oauth"))
+fitbit = LocalProxy(lambda: g.flask_dance_fitbit)