+++ flask_dance/contrib/spotify.py
@@ -1,7 +1,5 @@
-from functools import partial
-
-from flask import _app_ctx_stack as stack
-from flask.globals import LocalProxy, _lookup_app_object
+from flask import g
+from werkzeug.local import LocalProxy
 
 from flask_dance.consumer import OAuth2ConsumerBlueprint
 
@@ -75,10 +73,9 @@ def make_spotify_blueprint(
 
     @spotify_bp.before_app_request
     def set_applocal_session():
-        ctx = stack.top
-        ctx.spotify_oauth = spotify_bp.session
+        g.flask_dance_spotify = spotify_bp.session
 
     return spotify_bp
 
 
-spotify = LocalProxy(partial(_lookup_app_object, "spotify_oauth"))
+spotify = LocalProxy(lambda: g.flask_dance_spotify)